Subscribe working again.
Tricky here, we have a top level /entitlements resource in Candlepin,
represented by the Entitlement model in this code base. However in the
case of bind, we need to post to /consumers/UUID/entitlements, which is
difficult to navigate when trying to use active resource. Resorted to
using an explicit method to do this which calls some code used by Active
Resource custom methods.